[FIX] fieldservice_stock
