+++ fieldservice_stock/models/fsm_location.py
@@ -7,9 +7,9 @@ from odoo import api, fields, models
 class FSMLocation(models.Model):
     _inherit = 'fsm.location'
 
-    inventory_location_id = fields.Many2one('stock.location',
-                                            string='Inventory Location',
-                                            required=True)
+    inventory_location_id = fields.Many2one(
+        'stock.location', string='Inventory Location', required=True,
+        default=lambda self: self.env.ref('stock.stock_location_customers'))
     shipping_address_id = fields.Many2one('res.partner',
                                           string='Shipping Location')
