+++ fieldservice_stock/models/stock_request.py
@@ -44,7 +44,8 @@ class StockRequest(models.Model):
             vals['warehouse_id'] = fsm_order.warehouse_id.id
             val_date = vals['expected_date']
             if not isinstance(vals['expected_date'], str):
-                val_date = datetime.strftime(vals['expected_date'], '%Y-%m-%d %H:%M:%S')
+                val_date = datetime.strftime(vals['expected_date'],
+                                             '%Y-%m-%d %H:%M:%S')
             val_date = datetime.strptime(val_date, '%Y-%m-%d %H:%M:%S')
             picking_type_id = self.env['stock.picking.type'].search(
                 [('code', '=', 'stock_request_order'),
