+++ fieldservice_stock/models/stock_rule.py
@@ -1,0 +1,15 @@
+# Copyright (C) 2018 Brian McMaster
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+from odoo import models
+
+
+class StockRule(models.Model):
+    _inherit = 'stock.rule'
+
+    def _get_stock_move_values(self, product_id, product_qty, product_uom,
+                               location_id, name, origin, values, group_id):
+        vals = super()._get_stock_move_values(
+            product_id, product_qty, product_uom, location_id, name, origin,
+            values, group_id)
+        vals.update({'fsm_order_id': values.get('fsm_order_id')})
+        return vals
